+++ src/Beef.Core/RefData/ReferenceDataBaseGuid.cs
@@ -1,5 +1,6 @@
 ﻿// Copyright (c) Avanade. Licensed under the MIT License. See https://github.com/Avanade/Beef
 
+using Beef.Entities;
 using Newtonsoft.Json;
 using System;
 using System.Diagnostics;
@@ -11,7 +12,7 @@ namespace Beef.RefData
     /// </summary>
     [DebuggerDisplay("Id = {Id}, Code = {Code}, Text = {Text}, IsActive={IsActive}, IsValid={IsValid}")]
     [JsonObject(MemberSerialization.OptIn)]
-    public abstract class ReferenceDataBaseGuid : ReferenceDataBase
+    public abstract class ReferenceDataBaseGuid : ReferenceDataBase, IGuidIdentifier
     {
         /// <summary>
         /// Initializes a new instance of the <see cref="ReferenceDataBaseGuid"/> class.