+++ src/Beef.Core/Yaml/YamlConverter.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Avanade. Licensed under the MIT License. See https://github.com/Avanade/Beef
 
 using Beef.Entities;
+using Beef.RefData;
 using Newtonsoft.Json.Linq;
 using System;
 using System.Collections.Generic;
@@ -8,14 +9,14 @@ using System.IO;
 using System.Linq;
 using YamlDotNet.Serialization;
 
-namespace Beef.Test.NUnit.Internal
+namespace Beef.Yaml
 {
     /// <summary>
     /// Provides the capabilitity to convert <b>YAML</b> (or <b>JSON</b>) into a typed collection.
     /// </summary>
     public class YamlConverter
     {
-        private readonly JObject _json;
+        private readonly JArray _json;
 
         /// <summary>
         /// Gets or sets the default <see cref="ChangeLog.CreatedDate"/> value.
@@ -75,7 +76,7 @@ namespace Beef.Test.NUnit.Internal
         /// <param name="json">The <see cref="JObject"/> configuration.</param>
         private YamlConverter(JObject json)
         {
-            _json = json;
+            _json = json.Children().FirstOrDefault()?.Children().OfType<JArray>().FirstOrDefault();
         }
 
         /// <summary>
@@ -83,14 +84,17 @@ namespace Beef.Test.NUnit.Internal
         /// </summary>
         /// <param name="name">The name of the node that constains the items.</param>
         /// <param name="replaceAllShorthandGuids">Indicates whether to recursively replaces all '^n' values where 'n' is an integer with n.ToGuid() equivalent.</param>
-        /// <param name="setCreatedChangeLog">Indicates whether to set the <see cref="ChangeLog.CreatedBy"/> and <see cref="ChangeLog.CreatedDate"/> where <typeparamref name="T"/> implements <see cref="IChangeLog"/>.</param>
+        /// <param name="initializeCreatedChangeLog">Indicates whether to set the <see cref="ChangeLog.CreatedBy"/> and <see cref="ChangeLog.CreatedDate"/> where <typeparamref name="T"/> implements <see cref="IChangeLog"/>.</param>
+        /// <param name="initializeReferenceData">Indicates where to set the </param>
         /// <param name="itemAction">An action to allow further updating to each item.</param>
         /// <returns>The corresponding collection.</returns>
-        public IEnumerable<T> Convert<T>(string name = "items", bool replaceAllShorthandGuids = true, bool setCreatedChangeLog = true, Action<T> itemAction = null) where T : class, new()
+        public IEnumerable<T> Convert<T>(string name, bool replaceAllShorthandGuids = true, bool initializeCreatedChangeLog = true, bool initializeReferenceData = true, Action<JObject, int, T> itemAction = null) where T : class, new()
         {
-            var json = _json[Check.NotEmpty(name, nameof(name))];
+            Check.NotEmpty(name, nameof(name));
+
+            var json = _json?.Children().OfType<JObject>()?.Children().OfType<JProperty>().Where(x => x.Name == name)?.FirstOrDefault()?.Value;
             if (json == null || json.Type != JTokenType.Array)
-                throw new ArgumentException("Specified name either does not exist or is not an array.");
+                return default;
 
             if (replaceAllShorthandGuids)
                 ReplaceAllShorthandGuids(json);
@@ -99,7 +103,7 @@ namespace Beef.Test.NUnit.Internal
             foreach (var j in json.Children<JObject>())
             {
                 var val = j.ToObject<T>();
-                if (setCreatedChangeLog && val is IChangeLog cl)
+                if (initializeCreatedChangeLog && val is IChangeLog cl)
                 {
                     if (cl.ChangeLog == null)
                         cl.ChangeLog = new ChangeLog();
@@ -108,7 +112,10 @@ namespace Beef.Test.NUnit.Internal
                     cl.ChangeLog.CreatedDate = DefaultCreatedDate;
                 }
 
-                itemAction?.Invoke(val);
+                if (initializeReferenceData && val is ReferenceDataBase rdb)
+                    ReferenceDataInitializer(j, list.Count, rdb);
+
+                itemAction?.Invoke(j, list.Count, val);
 
                 list.Add(val);
             }
@@ -128,11 +135,45 @@ namespace Beef.Test.NUnit.Internal
                     var jv = (JValue)j;
                     var s = jv.Value as string;
                     if (!string.IsNullOrEmpty(s) && s.StartsWith("^") && int.TryParse(s.Substring(1), out var i))
-                        jv.Value = i.ToGuid();
+                        jv.Value = new Guid(i, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0);
                 }
                 else
                     ReplaceAllShorthandGuids(j);
             }
+        }
+
+        /// <summary>
+        /// Initialize reference data properties.
+        /// </summary>
+        private static void ReferenceDataInitializer(JObject json, int index, ReferenceDataBase value)
+        {
+            if (value.Code == null && json.Children().Count() == 1)
+            {
+                var jp = json.Children().OfType<JProperty>().First();
+                if (jp != null)
+                {
+                    value.Code = jp.Name;
+                    value.Text = jp.Value.ToString();
+                }
+            }
+
+            switch (value)
+            {
+                case ReferenceDataBaseGuid rdg:
+                    if (rdg.Id == Guid.Empty)
+                        rdg.Id = Guid.NewGuid();
+
+                    break;
+
+                case ReferenceDataBaseInt rdi:
+                    if (rdi.Id == 0)
+                        rdi.Id = index + 1;
+
+                    break;
+            }
+
+            if (value.SortOrder == 0)
+                value.SortOrder = index + 1;
         }
     }
 }